requirements.txt
fastapi==0.109.1
uvicorn[standard]==0.27.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
alembic==1.13.1
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator
google-generativeai
openai
httpx==0.26.0
uvloop==0.19.0
httptools==0.6.1